+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simplified "view instruction" modal
 - Included optional UI improvements (tour guide button color, tooltip for >10 instructions)
 
+- Added /polygon_operations endpoint with perform_polygon_operations for union, intersection, and difference of polygons, including validations.
+- Added Swagger UI documentation for the new polygon operations API.
+- Added pytest for polygon operation validation comprehensive test coverage.
 ### Changed
 - Replaced all usages of reactstrap and semantic-ui with CoreUI components and/or generic solutions 
 - moved quick_setup.py into main directory